Move cnl::is_composite into a bits/type_traits/
